CU-866a0pjt6 - Bug: banner shows wrong message
