+++ airbyte-webapp/src/views/layout/MainView/MainView.tsx
@@ -5,8 +5,8 @@ import { theme } from "theme";
 import { LoadingPage } from "components";
 import { CreateStepTypes } from "components/ConnectionStep";
 
-// import { useUser } from "core/AuthContext";
-// import { useAuthDetail } from "services/auth/AuthSpecificationService";
+import { useUser } from "core/AuthContext";
+import { getPaymentStatus, PAYMENT_STATUS } from "core/Constants/statuses";
 import { useHealth } from "hooks/services/Health";
 import useRouter from "hooks/useRouter";
 import { RoutePaths } from "pages/routePaths";
@@ -45,24 +45,11 @@ const hasCurrentStep = (state: unknown): state is { currentStep: string } => {
 const MainView: React.FC = (props) => {
   const { healthData } = useHealth();
   const { usage } = healthData;
-  // const { status } = useAuthDetail();
-  // const { user, updateUserStatus } = useUser();
+  const { user } = useUser();
   const { pathname, location, push } = useRouter();
   const [usagePercentage, setUsagePercentage] = useState<number>(0);
   const [isSidebar, setIsSidebar] = useState<boolean>(true);
   const [backgroundColor, setBackgroundColor] = useState<string>(theme.backgroundColor);
-
-  // useEffect(() => {
-  //   console.log("TEST: 01 PASS");
-  //   if (status && user.status !== status) {
-  //     console.log("TEST: 02 PASS");
-  //     console.log("====> STATUS <====");
-  //     console.log(status);
-  //     console.log("====> USER.STATUS <====");
-  //     console.log(user.status);
-  //     updateUserStatus?.(status);
-  //   }
-  // }, [status, updateUserStatus, user.status]);
 
   useEffect(() => {
     if (usage) {
@@ -136,13 +123,23 @@ const MainView: React.FC = (props) => {
     push(`/${RoutePaths.Settings}/${SettingsRoute.PlanAndBilling}`);
   };
 
+  const isUpgradePlanBanner = (): boolean => {
+    let showUpgradePlanBanner = false;
+    if (getPaymentStatus(user.status) === PAYMENT_STATUS.Free_Trial) {
+      if (!pathname.split("/").includes(RoutePaths.Payment)) {
+        showUpgradePlanBanner = true;
+      }
+    }
+    return showUpgradePlanBanner;
+  };
+
   return (
     <MainContainer>
       {isSidebar && <SideBar />}
       <Content backgroundColor={backgroundColor}>
         <ResourceNotFoundErrorBoundary errorComponent={<StartOverErrorView />}>
           <React.Suspense fallback={<LoadingPage />}>
-            <UpgradePlanBanner onBillingPage={onBillingPage} />
+            {isUpgradePlanBanner() && <UpgradePlanBanner onBillingPage={onBillingPage} />}
             {usage !== undefined && usagePercentage > 0 && usagePercentage < 100 && (
               <SyncNotificationBanner usagePercentage={usagePercentage} onBillingPage={onBillingPage} />
             )}
